LayoutTests/imported/w3c/web-platform-tests/event-timing/disconnect-target-expected.txt

ClickMe

FAIL Event Timing: when target is disconnected, entry.target returns null. assert_equals: There should be a single first-input entry expected 1 but got 0
